requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
feedgen==1.0.0
greenlet==3.2.1
html2text==2025.4.15
idna==3.10
lxml==5.3.2
playwright==1.52.0
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
